refactor: ProjectTable
