+++ components/ui/tables/TableHead.tsx
@@ -1,0 +1,12 @@
+export default function TableHead() {
+  return (
+    <thead>
+      <tr>
+        <th className="border px-4 py-2 text-left">ID</th>
+        <th className="border px-4 py-2 text-left">Title</th>
+        <th className="border px-4 py-2 text-left">Description</th>
+        <th className="border px-4 py-2 text-left">Tag</th>
+      </tr>
+    </thead>
+  );
+}
